refactor(history): refactor History page and create HistoryList component
